MOM6: +Made MOM_control_struct opaque
  Updated MOM6_examples to use a series of commits that eliminate the direct use
of elements of the MOM_control_struct at the driver level, and introduce a
shared MOM_state_type.  This set of commits include changes to the order of
entries in the MOM_parameter_doc files. In a note of caution, this sequence of
commit includes changes to mct_driver/ocn_comp_mct.F90 that mirror changes to
coupled_driver/ocean_model_MOM.F90, but have not been tested even to the point
of compiling due to the unavailability at GFDL of various software used by this
coupler; help is needed from someone who uses the mct_coupler to verify that
these have been done correctly.

- NOAA-GFDL/MOM6@b8c358e +Made MOM_control_struct opaque
- NOAA-GFDL/MOM6@93656d4 +Moved write_energy into step_MOM
- NOAA-GFDL/MOM6@dd8fdd6 Generate fatal error with excessive energy or NaNs
- NOAA-GFDL/MOM6@add27a3 Merge branch 'dev/gfdl' into restructure_MOM_CS
- NOAA-GFDL/MOM6@78031f9 +Moved ENERGYSAVEDAYS into MOM_sum_output.F90
- NOAA-GFDL/MOM6@a1bcc6f +Moved restart_CS up to the driver level
- NOAA-GFDL/MOM6@4747b1d +Reduced use of MOM_CSp elements by drivers
- NOAA-GFDL/MOM6@7b42449 Merge branch 'dev/gfdl' into restructure_MOM_CS
- NOAA-GFDL/MOM6@3cd24bc +Created a MOM_state_type