Use unit scaling type
  Added a new unit_scale_type and related module to handle dimensional unit
rescaling, and use it throughout the MOM6 code for rescaling Z_to_m and m_to_Z.
This involves numerous additional arguments and widespread changes.  There are
two new runtime debugging parameters and a third moved from the MOM_verticalGrid
module.  All answers are bitwise identical, but there are updates to all of the
MOM_parameter_doc files.  The list of commits in this PR include:
- NOAA-GFDL/MOM6@e1136e0 Rescale HMIX_SFC_PROP via get_param
- NOAA-GFDL/MOM6@fa04b52 Corrected openMP directives
- NOAA-GFDL/MOM6@04069b7 +Pass unit_scale_type argument to step_MOM_thermo
- NOAA-GFDL/MOM6@f8384ef +Replace GV%Z_to_m with US%Z_to_m
- NOAA-GFDL/MOM6@533d495 +Replace GV%m_to_Z with US%m_to_Z
- NOAA-GFDL/MOM6@9b542c3 +Add MOM_unit_scaling module