MOM6: +(*)Added MIN_SALINITY and fixed subML_N2
  Added a new run-time parameter to set the minimum salinity, corrected the
calculation of the diagnostic subML_N2 and corrected spelling error in
descriptions that are logged to the MOM_parameter_doc files.  By default, all
answers are bitwise identical, but there are numerous minor changes to the
MOM_parameter_doc files, which are being updated in MOM6-examples.  MOM6 commits
in this PR include:

 - NOAA-GFDL/MOM6@d927495 +Corrected spelling errors in documentation
 - NOAA-GFDL/MOM6@550308c +Corrected documentation of HALF_STRAT_DEPTH
 - NOAA-GFDL/MOM6@62c113e +Added MIN_SALINITY as a new runtime parameter.
 - NOAA-GFDL/MOM6@dc35d3a +(*)Revised the diagnostic subML_N2
 - NOAA-GFDL/MOM6@2015e63 (*)Revised propagate_int_tide code for symmetry