MOM6: +Removed hard newlines in get_param calls
  Removed hard newlines from numerous get_param and log_param descriptions
throughout the MOM6 codebase, relying instead on the new automatic newline
capability in writeMessageAndDesc.  In the case of formatted lists of options,
the hard newlines were retained. This cleans up the code and gives more standard
messages and smaller parameter_doc files, but it changes the comments in every
MOM_parameter_doc and SIS_parameter_doc file. All answers are bitwise identical.
This also corrects some updates to MOM_parameter_doc.all files that were omitted
in a recent hand-merge.  The MOM6 commit in this PR is:

- NOAA-GFDL/MOM6@a236032 +Removed hard newlines in get_param calls