*Set KVML = 0.0 in 31 cases
  Set the KVML enhancement to the near-surface viscosity to 0.0 m2/s in 31 cases
that do not need it to run, either because they are already using a physically-
based surface mixed layer scheme (like ePBL or KPP) to set the viscosity, or
because they have zero surface wind stress.  Also explicitly set KVML to the
current default values in another 8 MOM_input files, in preparation for
changing the default for KVML (or its renamed successor) to 0.  All cases are
qualitatively similar to before (apart from the ISOMIP runs, which show clear
improvements in their ability to adjust toward a reasonable state), but changing
the value of KVML does change answers.